third_party/blink/web_tests/platform/generic/fast/layout/common-ancestor-relayout-boundary-expected.txt

PASS internals.needsLayoutCount() is 4
PASS internals.needsLayoutCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
